requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.14
google-generativeai==1.0.0
python-dotenv==1.0.0
